+++ api/routes/recipes.js
@@ -11,7 +11,7 @@ router.post("/", RecipesController.create);
 // patch request most suitable - https://www.geeksforgeeks.org/what-is-the-difference-between-put-post-and-patch-in-restful-api/
 router.patch("/:recipe_id", RecipesController.updateRecipe);
 
-router.patch("/favouritedByOwner", tokenChecker, RecipesController.isFavourite);
+router.patch("/favouritedByOwner", RecipesController.isFavourite);
 
 module.exports = router;
